dlsep

+++ pipeline/volunteer_pipeline.sh
@@ -2,6 +2,7 @@
 #sra2mx for docker image
 #Copyright Mark Ziemann 2015 to 2017 mark.ziemann@gmail.com
 
+
 #Fix locale issue
 export LANGUAGE=C
 export LC_ALL=C
@@ -53,19 +54,28 @@ export -f exit1
 #JOB
 ORG=$1
 
-if [ $2 != '-f' ] ; then
-  SRR_FILE=$2
-  SRR=$(basename $SRR_FILE .sra)
+# is the download done separately?
+DL=FALSE
+if [ $2 == '-d' ] ; then
+  DL=TRUE
+  SRA_FILE=$3
+  SRR=$(echo $SRA_FILE | cut -d '_' -f2 | cut -d '.' -f1)
   echo $SRR
-  wget -O $SRR.html "https://www.ncbi.nlm.nih.gov/sra/${SRR}"
-  ORG2=$(echo $ORG | cut -c2-)
-  ORG_OK=$(sed 's/class=/\n/g' $SRR.html  | grep 'Organism:' | grep -c $ORG2)
-  rm $SRR.html
-  if [ $ORG_OK -ne 1 ] ; then
-    echo Annotated species name from NCBI SRA does not match user input! Quitting. | tee -a $SRR.log
-    exit1 ; exit 1
-  else
-    echo User input species and SRA metadata match. OK.
+else
+  if [ $2 != '-f' ] ; then
+    SRR_FILE=$2
+    SRR=$(basename $SRR_FILE .sra)
+    echo $SRR
+    wget -O $SRR.html "https://www.ncbi.nlm.nih.gov/sra/${SRR}"
+    ORG2=$(echo $ORG | cut -c2-)
+    ORG_OK=$(sed 's/class=/\n/g' $SRR.html  | grep 'Organism:' | grep -c $ORG2)
+    rm $SRR.html
+    if [ $ORG_OK -ne 1 ] ; then
+      echo Annotated species name from NCBI SRA does not match user input! Quitting. | tee -a $SRR.log
+      exit1 ; exit 1
+    else
+      echo User input species and SRA metadata match. OK.
+    fi
   fi
 fi
 
@@ -73,7 +83,7 @@ fi
 DEE_DIR=/dee2
 cd $DEE_DIR
 CODE_DIR=$DEE_DIR/code
-PIPELINE=$0
+PIPELINE=$CODE_DIR/volunteer_pipeline.sh
 PIPELINE_MD5=$(md5sum $PIPELINE | cut -d ' ' -f1)
 SW_DIR=$DEE_DIR/sw
 PATH=$PATH:$SW_DIR
@@ -312,6 +322,12 @@ if [ $2 != '-f' ] ; then
 ##########################################################################
   echo $SRR check if SRA file exists and download if neccessary
 ##########################################################################
+
+  # Moving SRA file downloaded previously
+  if [ $DL == 'TRUE' ] ; then
+    mv $SRA_FILE $SRR.sra
+  fi
+
   if [ ! -f $SRR.sra ] ; then
     #build URL
     BASEURL=ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra
@@ -1254,15 +1270,11 @@ MEM=$(echo $(free | awk '$1 ~ /Mem:/  {print $2-$3}') \
   $(free | awk '$1 ~ /Swap:/  {print $2-$3}') \
   | awk '{print $1+$2}' )
 NUM_CPUS=$(grep -c ^processor /proc/cpuinfo)
-CPU_SPEED=$(lscpu | grep MHz | awk '{print $NF}' | sort -k2gr)
+CPU_SPEED=$(lscpu | grep MHz | awk '{print $NF}' | head -1)
 
 ACC_URL="http://dee2.io/acc.html"
 ACC_REQUEST="http://dee2.io/cgi-bin/acc.sh"
-TMPHTML=/tmp/tmp.$RANDOM.html
-wget --no-check-certificate -r -O $TMPHTML "dee2.io/ip"
-SFTP_URL=$(cat $TMPHTML )
-
-rm $TMPHTML
+SFTP_URL=$(curl dee2.io/ip)
 
 if [ ! -z $MY_ORG ] ; then
   ORG_CHECK=$(echo 'athaliana celegans dmelanogaster drerio ecoli hsapiens mmusculus rnorvegicus scerevisiae' \
@@ -1289,25 +1301,11 @@ scerevisiae     1644684' | grep -w $MY_ORG | awk -v f=$MEM_FACTOR '{print $2*f}'
 fi
 
 if [ -z $MY_ORG ] ; then
-  ORGS=$(echo 'athaliana	2853904
-celegans	2652204
-dmelanogaster	3403644
-drerio	14616592
-ecoli	1576132
-hsapiens	28968508
-mmusculus	26069664
-rnorvegicus	26913880
-scerevisiae	1644684' | awk -v M=$MEM -v F=$MEM_FACTOR 'M>($2*F)' | sort -k2gr | awk '{print $1}')
 
-  TMPHTML=/tmp/tmp.$RANDOM.html
-  wget --no-check-certificate -O $TMPHTML "$ACC_URL"
-  wget --no-check-certificate -O $TMPHTML $(grep 'frame src=' $TMPHTML | cut -d '"' -f2)
+  echo "Error: no organism was specified. Quitting"
+  sleep 10
+  exit1
 
-  #specify organism if it has not already been specified by user
-  MY_ORG=$(join -1 1 -2 1 \
-  <(grep ORG $TMPHTML | cut -d '>' -f2 | tr -d ' .' | tr 'A-Z' 'a-z' | tr '()' ' ' | sort -k 1b,1) \
-  <(echo $ORGS | tr ' ' '\n' | sort -k 1b,1) | sort -k2gr | awk 'NR==1{print $1}' )
-  rm $TMPHTML
 fi
 
 #echo $MY_ORG
@@ -1315,22 +1313,17 @@ fi
 myfunc(){
 MY_ORG=$1
 ACC_REQUEST=$2
-TMPHTML=/tmp/tmp.$RANDOM.html
-wget --no-check-certificate -r -O $TMPHTML "${ACC_REQUEST}?ORG=${MY_ORG}&Submit"
-wget --no-check-certificate -O $TMPHTML $(grep 'frame src=' $TMPHTML | cut -d '"' -f2)
-ACCESSION=$(grep 'ACCESSION=' $TMPHTML | cut -d '=' -f2)
+ACCESSION=$(curl "http://dee2.io/cgi-bin/acc.sh?ORG=${MY_ORG}&submit" \
+| grep ACCESSION \
+| cut -d '=' -f2 )
+
 STAR --genomeLoad LoadAndExit --genomeDir ../ref/$MY_ORG/ensembl/star >/dev/null  2>&1
 echo $ACCESSION
-rm $TMPHTML
 }
 export -f myfunc
 
 key_setup(){
-TMPHTML=/tmp/tmp.$RANDOM.html
-wget --no-check-certificate -r -O $TMPHTML "dee2.io/ip"
-SFTP_URL=$(cat $TMPHTML )
 mkdir -p /dee2/.ssh
-rm $TMPHTML
 touch /dee2/.ssh/known_hosts
 chmod 777 /dee2/.ssh/known_hosts
 
@@ -1491,9 +1484,31 @@ else
 
 
 ##################################################
+# Testing whether the user is using the separate download process
+##################################################
+
+  DLPARAM=$(echo $@ | grep -wc '\-d')
+  if [ $DLPARAM -gt 0 ] ; then
+    FILECNT=$(ls /dee2/mnt/${MY_ORG}*.sra | wc -l)
+    while [ $FILECNT -gt 0 ] ; do
+      for SRA_FILE in /dee2/mnt/${MY_ORG}*.sra ; do
+        if [ ! -r /dee2/mnt/$SRA_FILE.started ] ; then
+          DIR=$(pwd)
+          echo Starting pipeline with species $1 and SRA file $SRA_FILE
+          touch $SRA_FILE.started
+          USER_ACCESSION=$(echo $SRA_FILE | cut -d '_' -f2 | cut -d '.' -f1)
+          main $1 -d $SRA_FILE VERBOSE=$VERBOSE
+          cd /dee2/data/$MY_ORG
+          zip -r /dee2/mnt/$USER_ACCESSION.$MY_ORG.zip $USER_ACCESSION
+          FILECNT=$(ls /dee2/mnt/${MY_ORG}*.sra | wc -l)
+        fi
+      done
+    done
+    exit
+  elif [[ $NUMVARS -eq "2" && $OWN_DATA -eq "0" ]] ; then
+##################################################
 # Testing whether the user has provided SRR accessions
 ##################################################
-  if [[ $NUMVARS -eq "2" && $OWN_DATA -eq "0" ]] ; then
     TESTACCESSIONS=$(echo $2 | tr ',' '\n' | cut -c2-3 | grep -vc RR)
     if [ $TESTACCESSIONS -eq 0 ] ; then
       for USER_ACCESSION in $(echo $2 | tr ',' ' ') ; do
